+++ ta_peersupervision/lib/pages/jadwal_page/psjadwal_page.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: use_build_context_synchronously
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
@@ -57,7 +55,7 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
         jadwal = fetchedEvents;
       });
     } catch (e) {
-   //   print('Failed to fetch events: $e');
+      // print('Failed to fetch events: $e');
       Get.snackbar('Jadwal Pendampingan', 'Belum ada Jadwal');
     }
   }
@@ -65,34 +63,28 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
   @override
   void dispose() {
     reqidController.dispose();
+    mediaController.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-   return LayoutBuilder(
+    return LayoutBuilder(
       builder: (context, constraints) {
         return Scaffold(
           key: scaffoldKey,
           backgroundColor: CustomColor.purpleBg,
-          endDrawer: constraints.maxWidth >= minDesktopWidth
-          ? null
-          : const PSDrawerMobile(),
-          
+          endDrawer: constraints.maxWidth >= minDesktopWidth ? null : const PSDrawerMobile(),
           body: SingleChildScrollView(
             controller: scrollController,
             scrollDirection: Axis.vertical,
-
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 if (constraints.maxWidth >= minDesktopWidth)
-                // Main Container
-                  HeaderPSBack(onNavMenuTap: (int navIndex){
-                    // call function
+                  HeaderPSBack(onNavMenuTap: (int navIndex) {
                     scrollToSection(navIndex);
-                  },)
-
+                  })
                 else
                   PSHeaderMobile(
                     onLogoTap: () {},
@@ -100,7 +92,6 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
                       scaffoldKey.currentState?.openEndDrawer();
                     },
                   ),
-
                 const SizedBox(height: 30),
                 CalendarWidget(
                   onDaySelected: _showEventDialog,
@@ -119,34 +110,39 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
 
   void _showEventDialog(DateTime date, List<MyJadwal> events) {
     DateTime selectedDate = DateTime(date.year, date.month, date.day);
-//    print(selectedDate);
+    DateTime today = DateTime.now();
+    bool isPastDate = selectedDate.isBefore(DateTime(today.year, today.month, today.day));
 
     showDialog(
       context: context,
       builder: (BuildContext context) {
         return AlertDialog(
-          title: Text('Tambahkan Jadwal Pendampingan ${date.day}/${date.month}/${date.year}'),
+          title: Text(isPastDate
+              ? 'Jadwal Pendampingan ${date.day}/${date.month}/${date.year}'
+              : 'Tambahkan Jadwal Pendampingan ${date.day}/${date.month}/${date.year}'),
           content: SingleChildScrollView(
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                TextField(
-                  controller: reqidController,
-                  decoration: const InputDecoration(labelText: 'ID Dampingan'),
-                  keyboardType: TextInputType.number,
-                  inputFormatters: <TextInputFormatter>[
-                    FilteringTextInputFormatter.digitsOnly,
-                  ],
-                ),
-                TextField(
-                  controller: mediaController,
-                  decoration: const InputDecoration(labelText: 'Media Pendampingan'),
-                ),
+                if (!isPastDate) ...[
+                  TextField(
+                    controller: reqidController,
+                    decoration: const InputDecoration(labelText: 'ID Dampingan'),
+                    keyboardType: TextInputType.number,
+                    inputFormatters: <TextInputFormatter>[
+                      FilteringTextInputFormatter.digitsOnly,
+                    ],
+                  ),
+                  TextField(
+                    controller: mediaController,
+                    decoration: const InputDecoration(labelText: 'Tempat Pendampingan'),
+                  ),
+                ],
                 const SizedBox(height: 16),
                 const Text('Pendampingan Hari Ini:', style: TextStyle(fontWeight: FontWeight.bold)),
                 const SizedBox(height: 8),
                 if (events.isEmpty)
-                  const Text('Tidak ada pendampingan di hari ini')
+                  const Text('Tidak ada pendampingan di hari ini', style: TextStyle(color: Color.fromARGB(255, 227, 152, 147)))
                 else
                   Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
@@ -155,28 +151,29 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
                           Expanded(
-                            child: Text('${event.initial}\nID Dampingan: ${event.reqid}\nMedia Pendampingan: ${event.mediapendampingan}\n'),
+                            child: SelectableText('${event.initial}\nID Dampingan: ${event.reqid}\nTempat Pendampingan: ${event.mediapendampingan}\n'),
                           ),
-                          IconButton(
-                            icon: const Icon(Icons.delete),
-                            onPressed: () async {
-                              final confirmDelete = await _showDeleteConfirmationDialog(context, event.jadwalid);
-                              if (confirmDelete) {
-                                // Menghapus event dari map terlebih dahulu
-                                setState(() {
-                                  jadwal[selectedDate]?.remove(event);
-                                  if (jadwal[selectedDate]?.isEmpty ?? false) {
-                                    jadwal.remove(selectedDate);
-                                  }
-                              });
-                              // Menghapus event dari repository
-                              await repository.deleteJadwal(event.jadwalid);
-                              // Mengambil ulang data dari repository
-                              await _fetchEvents(widget.psnim);
-                              Navigator.of(context).pop();
-                            }
-                            },
-                          ),
+                          if (!isPastDate)
+                            IconButton(
+                              icon: const Icon(Icons.delete),
+                              onPressed: () async {
+                                final confirmDelete = await _showDeleteConfirmationDialog(context, event.jadwalid);
+                                if (confirmDelete) {
+                                  // Menghapus event dari map terlebih dahulu
+                                  setState(() {
+                                    jadwal[selectedDate]?.remove(event);
+                                    if (jadwal[selectedDate]?.isEmpty ?? false) {
+                                      jadwal.remove(selectedDate);
+                                    }
+                                  });
+                                  // Menghapus event dari repository
+                                  await repository.deleteJadwal(event.jadwalid);
+                                  // Mengambil ulang data dari repository
+                                  await _fetchEvents(widget.psnim);
+                                  Navigator.of(context).pop();
+                                }
+                              },
+                            ),
                         ],
                       );
                     }).toList(),
@@ -184,37 +181,45 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
               ],
             ),
           ),
-          actions: <Widget>[
-            ElevatedButton(
-              onPressed: () {
-                if (mediaController.text.isEmpty || reqidController.text.isEmpty) {
-                  Get.snackbar('Rencanakan jadwal Pendampingan', 'Kolom harus terisi!',
-                      backgroundColor: Colors.red,
-                      colorText: Colors.white);
-                } else {
-                  Jadwal jadwal = Jadwal(
-                    reqid: int.parse(reqidController.text),
-                    tanggal: formatDateSQL(date),
-                    mediapendampingan: mediaController.text,
-                  );
+          actions: isPastDate
+              ? <Widget>[
+                  TextButton(
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                    },
+                    child: const Text('Tutup'),
+                  ),
+                ]
+              : <Widget>[
+                  ElevatedButton(
+                    onPressed: () {
+                      if (mediaController.text.isEmpty || reqidController.text.isEmpty) {
+                        Get.snackbar('Rencanakan jadwal Pendampingan', 'Kolom harus terisi!',
+                            backgroundColor: Colors.red, colorText: Colors.white);
+                      } else {
+                        Jadwal jadwal = Jadwal(
+                          reqid: int.parse(reqidController.text),
+                          tanggal: formatDateSQL(date),
+                          mediapendampingan: mediaController.text,
+                        );
 
-                  repository.createJadwal(jadwal: jadwal).then((value) {
-                    _fetchEvents(widget.psnim);
-                    Navigator.of(context).pop();
-                  });
-                  reqidController.clear();
-                  mediaController.clear();
-                }
-              },
-              child: const Text('Simpan'),
-            ),
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: const Text('Tutup', style: TextStyle(color: Colors.red)),
-            ),
-          ],
+                        repository.createJadwal(jadwal: jadwal).then((value) {
+                          _fetchEvents(widget.psnim);
+                          Navigator.of(context).pop();
+                        });
+                        reqidController.clear();
+                        mediaController.clear();
+                      }
+                    },
+                    child: const Text('Simpan'),
+                  ),
+                  TextButton(
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                    },
+                    child: const Text('Tutup'),
+                  ),
+                ],
         );
       },
     );
@@ -229,15 +234,16 @@ class _PSJadwalPageState extends State<PSJadwalPage> {
           content: Text('Apakah Anda yakin ingin menghapus jadwal pendampingan dengan ID: $jadwalid?'),
           actions: [
             TextButton(
-              onPressed: () => Navigator.of(context).pop(false),
-              child: const Text('Batal'),
-            ),
-            TextButton(
               onPressed: () => Navigator.of(context).pop(true),
               style: ButtonStyle(
                 foregroundColor: MaterialStateProperty.all<Color>(Colors.red),
               ),
               child: const Text('Hapus'),
+            ),
+            const SizedBox(width: 10,),
+            TextButton(
+              onPressed: () => Navigator.of(context).pop(false),
+              child: const Text('Batal'),
             ),
           ],
         );
